Move sunweg sensor definitions
