+++ homeassistant/components/sunweg/sensor/inverter.py
