+++ homeassistant/components/sunweg/sensor/phase.py
