+++ homeassistant/components/sunweg/sensor/string.py
