+++ homeassistant/components/sunweg/sensor/total.py
