--- homeassistant/components/sunweg/sensor_types/__init__.py
@@ -1,1 +1,0 @@
-"""Sensor types for supported Sun WEG systems."""
