+++ tests/components/sunweg/test_init.py
@@ -7,7 +7,7 @@ from sunweg.api import APIHelper, SunWegApiError
 
 from homeassistant.components.sunweg import SunWEGData
 from homeassistant.components.sunweg.const import DOMAIN, DeviceType
-from homeassistant.components.sunweg.sensor_types.sensor_entity_description import (
+from homeassistant.components.sunweg.sensor.sensor_entity_description import (
     SunWEGSensorEntityDescription,
 )
 from homeassistant.config_entries import ConfigEntryState
